Add back to query test and split up assert
